Get completely connected topological case working
In this case, it is equivalent to using the global best all the time.

+++ shaders/update_topological_best_complete.frag
@@ -5,10 +5,10 @@ precision highp int;
 
 uniform sampler2D best_error_value_texture;
 
-layout (location = 0) out vec4 topological_best_idx;
+layout (location = 0) out uvec4 topological_best_idx;
 
 in vec2 cc;
 
 void main() {
-    topological_best_idx = vec4(texelFetch(best_error_value_texture, ivec2(0, 0), 0).yz, 0.0, 0.0);
+    topological_best_idx = uvec4(texelFetch(best_error_value_texture, ivec2(0, 0), 0).yz, 0.0, 0.0);
 }
